# [MRG] issue #13383: Made plot_compare_reduction.py faster