--- a/examples/compose/plot_compare_reduction.py
+++ b/examples/compose/plot_compare_reduction.py
@@ -29,7 +29,6 @@ fitting of a transformer is costly.
 
 # Authors: Robert McGibbon, Joel Nothman, Guillaume Lemaitre
 
-
 import numpy as np
 import matplotlib.pyplot as plt
 from sklearn.datasets import load_digits
@@ -40,6 +39,7 @@ from sklearn.decomposition import PCA, NMF
 from sklearn.feature_selection import SelectKBest, chi2
 
 print(__doc__)
+
 
 pipe = Pipeline([
     # the reduce_dim stage is populated by the param_grid
@@ -63,7 +63,7 @@ param_grid = [
 ]
 reducer_labels = ['PCA', 'NMF', 'KBest(chi2)']
 
-grid = GridSearchCV(pipe, cv=5, n_jobs=1, param_grid=param_grid)
+grid = GridSearchCV(pipe, cv=3, n_jobs=1, param_grid=param_grid)
 digits = load_digits()
 grid.fit(digits.data, digits.target)
 
@@ -113,7 +113,7 @@ cached_pipe = Pipeline([('reduce_dim', PCA()),
                        memory=memory)
 
 # This time, a cached pipeline will be used within the grid search
-grid = GridSearchCV(cached_pipe, cv=5, n_jobs=1, param_grid=param_grid)
+grid = GridSearchCV(cached_pipe, cv=3, n_jobs=1, param_grid=param_grid)
 digits = load_digits()
 grid.fit(digits.data, digits.target)
 
@@ -128,3 +128,4 @@ rmtree(location)
 # estimator data, leading to save processing time. Therefore, the use of
 # caching the pipeline using ``memory`` is highly beneficial when fitting
 # a transformer is costly.
+