@qinhanmin2014 2019-07-14
apologies the log in the issue seems outdated
in master, the example only takes 8 sec, this is also the case on my PC.