@amueller 2019-07-14
No harm in speeding it up, though?