@qinhanmin2014 2019-07-14
> No harm in speeding it up, though?

cv=5 -> cv=3 (not recommended now), 8s -> 5s (trivial improvement IMO). This is why I close this @amueller .
But if you like it, please reopen and merge.