@amueller 2019-07-14
You're right, it might be more important to show best practices for cv.